Move linkable contract to web3.contract
